+++ app/src/main/java/com/subhajitrajak/durare/data/repositories/WeightSetupRepository.kt
@@ -15,7 +15,7 @@ class WeightSetupRepository (val context: Context) {
     val userDoc = uid?.let { db.collection(Constants.USERS).document(it) }
     val profileDoc = userDoc?.collection(Constants.USER_DATA)?.document(Constants.PROFILE)
 
-    suspend fun saveData(weight: Float) {
+    suspend fun saveData(weight: Double) {
         pref.setWeight(weight)
         profileDoc?.update(Constants.USER_WEIGHT, weight)?.await()
     }
