+++ app/src/main/java/com/subhajitrajak/durare/utils/Preferences.kt
@@ -71,8 +71,8 @@ class Preferences private constructor(context: Context) {
     fun getGoal(): Int = prefs.getInt(GOAL, 50)
     fun setGoal(value: Int) = prefs.edit { putInt(GOAL, value) }
 
-    fun getWeight(): Float = prefs.getFloat(WEIGHT, 0f)
-    fun setWeight(value: Float) = prefs.edit { putFloat(WEIGHT, value) }
+    fun getWeight(): Double = prefs.getFloat(WEIGHT, 0f).toDouble()
+    fun setWeight(value: Double) = prefs.edit { putFloat(WEIGHT, value.toFloat()) }
 
     // reset all preferences in personalize screen to defaults
     fun resetPersonalizationsToDefaults() {
